Update token conversion after oneof removal, disable depguard lint
